Switch xt_app default UI to egui
